pyproject.toml
[project]
name = "reasondb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gdown>=5.2.0",
    "gpustat>=1.1.1",
    "librosa>=0.11.0",
    "pandasql>=0.7.3",
    "psutil>=7.1.3",
    "scikit-learn>=1.7.2",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "wheel>=0.45.1",
]
